Split entrypoint_lookup implementation and interface
This reorganizes code so that the interface and general code that uses
that interface is separated from the code that implements the interface
with a K8s-specific implementation.
